+++ protected/humhub/modules/topic/models/Topic.php
@@ -160,16 +160,18 @@ class Topic extends ContentTag
         }
     }
 
-    public static function isAllowedToCreate(ContentContainerActiveRecord $contentContainer)
+    public static function isAllowedToCreate(?ContentContainerActiveRecord $contentContainer = null): bool
     {
-        return (
-            $contentContainer instanceof Space
-            && Yii::$app->getModule('space')->settings->get('allowSpaceTopics', true)
-            && $contentContainer->can(AddTopic::class)
-        )
-        || (
-            $contentContainer instanceof User
-            && Yii::$app->getModule('user')->settings->get('auth.allowUserTopics', true)
-        );
+        if ($contentContainer instanceof Space) {
+            return Yii::$app->getModule('space')->settings->get('allowSpaceTopics', true)
+                && $contentContainer->can(AddTopic::class);
+        }
+
+        if ($contentContainer instanceof User) {
+            return Yii::$app->getModule('user')->settings->get('auth.allowUserTopics', true);
+        }
+
+        // Global Topics can be created only in Administration
+        return false;
     }
 }